docker: Move docker/containers out of the shell package.
