+++ pkg/shell/manifest.json
@@ -19,6 +19,7 @@
             "label": "Networking",
             "order": 10
         }
+
     },
 
     "tools": {
